add synchronizer for single set since the brickset api only populates some fields (description, tags, ...) when you call get single set explicitly! add tests for the above update ParameterUserHashSetId to use the same type for SetID as the underlying data persistence layer

+++ abremir.AllMyBricks.ThirdParty.Brickset/Models/ParameterUserHashSetId.cs
@@ -2,6 +2,6 @@
 {
     public class ParameterUserHashSetId : ParameterApiKeyUserHash
     {
-        public int SetID { get; set; }
+        public long SetID { get; set; }
     }
 }
